break out styles and types
