+++ packages/badge/src/index.ts
@@ -1,3 +1,3 @@
 import Badge from './Badge';
-export { Variant } from './Badge';
+export { Variant } from './types';
 export default Badge;
